fix: sidebar for explorer page
